app/requirements.txt
python-dotenv
python-crontab
telethon
